[Refactor] submit 함수 리팩토링
